% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_model_posteriors.R
\name{get_meta_model_posteriors}
\alias{get_meta_model_posteriors}
\title{Fit meta models and return posterior draws}
\usage{
get_meta_model_posteriors(
  data = NULL,
  n_posterior_draws = 100,
  suppressor = suppress_mwo
)
}
\arguments{
\item{data}{simulated data}

\item{n_posterior_draws}{number of posterior draws after warm-up}
}
\value{
List of posterior draws for each model
}
\description{
Fit meta models and return posterior draws
}
