% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level3.R
\name{level_3}
\alias{level_3}
\title{Obtain meta_uncertainty_fit object given observed and simulated posterior model probability.}
\usage{
level_3(pmp_obs, pmp_sim, n_posterior_draws = 1000, eps = 1e-10, verbosity = 1)
}
\arguments{
\item{pmp_obs}{vector of observed model probability.}

\item{pmp_sim}{dataframe containing index of true model in the first column and the pmp corresponding to each models.}

\item{n_posterior_draws}{number of posterior draws after warm-up}

\item{eps}{small number to be added / subtracted when posterior model probability takes 0 / 1}
}
\value{
meta_uncertainty object
}
\description{
Obtain meta_uncertainty_fit object given observed and simulated posterior model probability.
}
