% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_meta_model_density}
\alias{plot_meta_model_density}
\title{Plot simulated posterior model probability in ternary plot.}
\usage{
plot_meta_model_density(metabmcfit)
}
\arguments{
\item{metabmcfit}{metabmcfit object created by \code{\link{metabmc}}.  Only the fit with three models is supported.}
}
\value{
Ternary plot
}
\description{
Plot simulated posterior model probability in ternary plot.
}
