% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_predictive_mixture}
\alias{plot_predictive_mixture}
\title{Plot simulated posterior model probability in ternary plot.}
\usage{
plot_predictive_mixture(metabmcfit)
}
\arguments{
\item{metabmcfit}{metabmcfit object created by \code{\link{metabmc}}. Only the fit with three models is supported.}
}
\value{
Ternary plot
}
\description{
Plot simulated posterior model probability in ternary plot.
}
