% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabmc.R
\name{metabmc}
\alias{metabmc}
\title{Fit meta model for posterior model probabilities and predictive mixture model}
\usage{
metabmc(
  ...,
  formula_list,
  prior_list,
  family_list = rep(NULL, 3),
  data = NULL,
  n_sim = 20,
  n_posterior_draws = 1000,
  warmup = 5000,
  prior_model_prob = "uniform",
  eps = 0.001,
  brms_arg_list = NULL,
  verbosity = 1
)
}
\arguments{
\item{...}{brmsfit object}

\item{formula_list}{List of brmsformula}

\item{prior_list}{List of vector of prior}

\item{family_list}{List of brmsfamily}

\item{data}{Data to fit}

\item{n_sim}{Number of simulation to obtain posterior model probability}

\item{n_posterior_draws}{Number of posterior draws}

\item{warmup}{Number of warmup}

\item{prior_model_prob}{Prior distribution from which true model is drawn}

\item{eps}{Small number to be added / subtracted when posterior model probability takes 0 / 1}

\item{brms_arg_list}{Additional argument to be passed to brms::brm when fitting model}
}
\value{
metabmc object which is to be passed to plotting function for inspection of uncertainty in posterior model probability.
}
\description{
Fit meta model for posterior model probabilities and predictive mixture model.
All-in-one function to simulate posterior model probabilities given model specification and prior, estimate density
over simulated posterior model probabilities (meta models), and estimate predictive mixture model combining meta models
and observed posterior model probabilities. Currently only three model candidate is supported.
After fitting the models, results can be visualized using following three functions;
\code{\link{plot_simulated_pmp}}, \code{\link{plot_meta_model_density}}, and \code{\link{plot_predictive_mixture}}.
}
\examples{
\dontrun{
# brmsfit objects as arguments
fit1 <- brm(y ~ (1 | g) + x, data = d)
fit2 <- brm(y ~ x, data = d)
fit3 <- brm(y ~ (x | g), data = d)
metabmc_fit <- metabmc(fit1, fit2, fit3, n_sim=10)

# Using list of formula, data and other arguments
m1 <- bf(y ~ (1 | g) + x)
m2 <- bf(y ~ x)
m3 <- bf(y ~ (x | g))

metabmc_fit <- metabmc(list(m1, m2, m3), data=d, n_sim=10)
}
}
