% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabmc.R
\name{metabmc}
\alias{metabmc}
\title{Fit Meta Uncertainty model. Given brms}
\usage{
metabmc(
  ...,
  formula_list,
  prior_list,
  family_list = rep(NULL, 3),
  data = NULL,
  n_sim = 100,
  n_posterior_draws = 1000,
  warmup = 5000,
  prior_model_prob = "uniform",
  eps = 1e-05,
  brms_arg_list = NULL
)
}
\arguments{
\item{...}{brmsfit object}

\item{formula_list}{List of brmsformula}

\item{prior_list}{List of vector of prior}

\item{family_list}{List of brmsfamily}

\item{data}{Data to fit}

\item{n_sim}{Number of simulation to obtain posterior model probability}

\item{n_posterior_draws}{Number of posterior draws}

\item{warmup}{Number of warmup}

\item{prior_model_prob}{Prior distribution from which true model is drawn}

\item{eps}{Small number to be added / subtracted when posterior model probability takes 0 / 1}

\item{brms_arg_list}{Additional argument to be passed to brms::brm when fitting model}
}
\value{
metabmc object which is to be passed to plotting function for inspection of uncertainty in posterior model probability.
}
\description{
Fit Meta Uncertainty model. Given brms
}
