% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_meta_model_density}
\alias{plot_meta_model_density}
\title{Plot density of metamodels}
\usage{
plot_meta_model_density(metabmcfit)
}
\arguments{
\item{metabmcfit}{metabmcfit object created by \code{\link{metabmc}}.  Only the fit with three models is supported.}
}
\value{
Three triangle plot of density of meta models (Distribution of posterior model probability)
}
\description{
Plot density of metamodels
}
