% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_predictive_mixture}
\alias{plot_predictive_mixture}
\title{Plot predictive mixture model}
\usage{
plot_predictive_mixture(metabmcfit)
}
\arguments{
\item{metabmcfit}{metabmcfit object created by \code{\link{metabmc}}. Only the fit with three models is supported.}
}
\value{
Single ternary plot of density of predictive mixture model along with observed PMPs.
}
\description{
Plot predictive mixture model
}
\examples{
\dontrun{
plot_predictive_mixture(metabmc_fit)
}
}
