% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_simulated_pmp}
\alias{plot_simulated_pmp}
\title{Plot simulated posterior model probability.}
\usage{
plot_simulated_pmp(metabmcfit)
}
\arguments{
\item{metabmcfit}{metabmcfit object created by \code{\link{metabmc}}. Only the fit with three models is supported.}
}
\value{
Three triangle plot showing simulated posterior model probability from three true models.
}
\description{
Plot simulated posterior model probability.
}
